+++ src/platform/node/extension-manager/handler.ts
@@ -127,8 +127,8 @@ export class ModuleHandler extends CoreHandler {
   }
   private executeByAppNpm(command: string, modules: any[], resolve, reject) {
     // https://www.npmjs.com/package/bin-links
-    npmCli.load({ 'bin-links': false, verbose: true, prefix: this.baseDir }, loaderr => {
-      const moduleList = modules.map(({ name, version }) => (version ? `${name}@${version}` : name));
+    npmCli.load({ 'bin-links': false, verbose: true, prefix: this.baseDir, registry: this.registry }, loaderr => {
+      const moduleList = modules.filter(val => val).map(({ name, version }) => (version ? `${name}@${version}` : name));
       let executeCommand = ['update', 'install', 'uninstall'];
       if (!executeCommand.includes(command)) {
         return;
@@ -165,7 +165,7 @@ export class ModuleHandler extends CoreHandler {
       // this.executeBySystemNpm(command, modules, resolve)
       // * Set Proxy
       // * Set registry
-      await this.setRegistry();
+      // await this.setRegistry();
       this.executeByAppNpm(command, modules, resolve, reject);
     });
   }