+++ src/workbench/browser/src/app/modules/extension-select/export-api/export-api.component.ts
@@ -36,7 +36,7 @@ export class ExportApiComponent implements OnInit {
       .get()
       .pipe(takeUntil(this.destroy$))
       .subscribe((inArg: Message) => {
-        if (inArg.type === 'installedExtensionsChange') {
+        if (inArg.type === 'extensionsChange') {
           this.initData();
         }
       });