+++ src/workbench/browser/src/app/pages/pages.module.ts
@@ -9,13 +9,19 @@ import { ToolbarModule } from '../layouts/toolbar/toolbar.module';
 import { NpsMaskModule } from '../modules/nps-mask/nps-mask.module';
 import { SystemSettingModule } from '../modules/system-setting/system-setting.module';
 import { SharedModule } from '../shared/shared.module';
+import { ChatgptRobotComponent } from './components/chatgpt-robot/chatgpt-robot.component';
 import { ThirdLoginComponent } from './components/third-login/third-login.component';
 import { UserModalComponent } from './components/user-modal/user-modal.component';
 import { PagesRoutingModule } from './pages-routing.module';
 import { PagesComponent } from './pages.component';
 
 @NgModule({
+  declarations: [PagesComponent, SidebarComponent, LocalWorkspaceTipComponent, UserModalComponent, ThirdLoginComponent],
+  exports: [],
+  providers: [],
+  schemas: [],
   imports: [
+    ChatgptRobotComponent,
     PagesRoutingModule,
     NpsMaskModule,
     NzNotificationModule,
@@ -24,10 +30,6 @@ import { PagesComponent } from './pages.component';
     NavbarModule,
     ToolbarModule,
     SharedModule
-  ],
-  declarations: [PagesComponent, SidebarComponent, LocalWorkspaceTipComponent, UserModalComponent, ThirdLoginComponent],
-  exports: [],
-  providers: [],
-  schemas: []
+  ]
 })
 export class PagesModule {}