+++ src/workbench/browser/src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
-import { APP_CONFIG_INSTANT, COMMON_CONFIG } from 'eo/workbench/browser/src/environments/common';
+import { APP_CONFIG_INSTANT, COMMON_CONFIG } from 'eo/workbench/browser/src/environments/common.constant';
 
 export const APP_CONFIG = {
   serverUrl: 'https://postcat.com',